backend/requirements.txt
# ============================================================
# LLM 智能客服系统 - Python 依赖
# ============================================================

# --- Web 框架 ---
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# --- LLM & AI ---
langchain==0.3.7
langchain-openai==0.2.7
langchain-community==0.3.7
langchain-chroma==0.1.4
openai==1.54.0
tiktoken==0.8.0

# --- 向量数据库 ---
chromadb==0.5.18

# --- 数据库 ---
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# --- 缓存 ---
redis==5.2.0
hiredis==3.0.0

# --- 认证 ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# --- 配置管理 ---
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0

# --- 文档处理 ---
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# --- 数据处理 ---
numpy==1.26.4
pandas==2.2.3

# --- HTTP 客户端 ---
httpx==0.27.2
aiohttp==3.11.0

# --- 工具 ---
loguru==0.7.2
tenacity==9.0.0
python-slugify==8.0.4

# --- 测试 ---
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
factory-boy==3.3.1
faker==33.0.0
